lib: fwts_checksum: object sizes should be size_t rather than int
Checksum lengths should be on size_t sizes rather than ints.

Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>

+++ src/lib/include/fwts_checksum.h
@@ -22,6 +22,6 @@
 
 #include "fwts.h"
 
-uint8_t fwts_checksum(const uint8_t *data, const int length);
+uint8_t fwts_checksum(const uint8_t *data, const size_t length);
 
 #endif
